refactor(frontend): migrate settings surface into features
