+++ src/frontend/src/routes/app/settings.tsx
@@ -2,7 +2,7 @@ import { createFileRoute, lazyRouteComponent } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/app/settings")({
   component: lazyRouteComponent(
-    () => import("@/screens/settings/settings-screen"),
+    () => import("@/features/settings/settings-screen"),
     "SettingsScreen",
   ),
 });
