+++ src/frontend/vite.config.ts
@@ -221,8 +221,8 @@ export default defineConfig({
         "src/features/workspace/workspace-screen.tsx",
         "src/app/workspace/transcript/workspace-message-list.tsx",
         "src/app/workspace/workspace-composer.tsx",
-        "src/screens/settings/settings-screen.tsx",
-        "src/screens/settings/runtime-form.tsx",
+        "src/features/settings/settings-screen.tsx",
+        "src/features/settings/runtime-form.tsx",
         "src/features/volumes/volumes-screen.tsx",
         "src/features/volumes/volumes-canvas-panel.tsx",
       ],
